+++ src/main/java/clank/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package clank.command;
+
+import clank.task.TaskList;
+import clank.utility.Ui;
+import clank.utility.Storage;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
