+++ src/main/java/clank/exception/ClankException.java
@@ -1,3 +1,5 @@
+package clank.exception;
+
 public class ClankException extends RuntimeException {
     public ClankException(String message) {
         super(message);
