+++ src/main/java/clank/exception/EmptyDescriptionException.java
@@ -1,3 +1,5 @@
+package clank.exception;
+
 public class EmptyDescriptionException extends ClankException {
     public EmptyDescriptionException(String command) {
         super("Oops! You have to specify the description for \"" + command +
