+++ src/main/java/clank/exception/InvalidFormatException.java
@@ -1,3 +1,5 @@
+package clank.exception;
+
 public class InvalidFormatException extends ClankException {
     public InvalidFormatException() {
       super("Oh no! You've entered an invalid index! Please enter a valid number!");
