+++ src/main/java/clank/task/Event.java
@@ -1,20 +1,26 @@
+package clank.task;
+
+import clank.exception.ClankException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task {
+public class Event extends Task {
     private static final DateTimeFormatter INPUT_FORMATTER = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
     private static final DateTimeFormatter OUTPUT_FORMATTER =
             DateTimeFormatter.ofPattern("d MMMM yyyy, h:mm a");
 
-    protected LocalDateTime by;
+    protected LocalDateTime start;
+    protected LocalDateTime end;
 
-    public Deadline(String description, String by) {
+    public Event(String description, String start, String end) {
         super(description);
         try {
-            this.by = parseDateTime(by);
+            this.start = parseDateTime(start);
+            this.end = parseDateTime(end);
         } catch (DateTimeParseException e) {
-            throw new ClankException("Error parsing deadline");
+            throw new ClankException("Error parsing event");
         }
     }
 
@@ -24,11 +30,13 @@ public class Deadline extends Task {
 
     @Override
     public String toSaveFormat() {
-        return "D|" + isDone + "|" + description + "|" + by.format(INPUT_FORMATTER);
+        return "T|" + isDone + "|" + description
+                + "|" + start.format(INPUT_FORMATTER) + "|" + end.format(INPUT_FORMATTER);
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " + by.format(OUTPUT_FORMATTER) + ")";
+        return "[E]" + super.toString() + " (from: " + start.format(OUTPUT_FORMATTER)
+                + " to: " + end.format(OUTPUT_FORMATTER) + ")";
     }
 }
