+++ src/main/java/clank/task/Task.java
@@ -1,3 +1,5 @@
+package clank.task;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
