+++ src/main/java/clank/task/Todo.java
@@ -1,3 +1,5 @@
+package clank.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
