+++ src/main/java/clank/utility/Parser.java
@@ -1,3 +1,8 @@
+package clank.utility;
+
+import clank.command.*;
+import clank.exception.UnknownCommandException;
+
 public class Parser {
     public static Command parse(String input) {
         String[] parts = input.split(" ", 2);
