+++ src/main/java/clank/utility/Ui.java
@@ -1,3 +1,5 @@
+package clank.utility;
+
 import java.util.Scanner;
 
 public class Ui {
@@ -8,7 +10,7 @@ public class Ui {
     }
 
     public void showWelcomeMessage() {
-        System.out.println("Greetings! I'm Clank, Ratchet's best friend.\n"
+        System.out.println("Greetings! I'm clank.Clank, Ratchet's best friend.\n"
                 + "I heard you are in trouble, and I'm here to help.\n"
                 + "How may I assist you?");
     }
