+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
- Greetings! I'm Clank, Ratchet's best friend.
+ Greetings! I'm clank.Clank, Ratchet's best friend.
  I heard you are in trouble, and I'm here to help.
  How may I assist you?
 ____________________________________________________________
